# docs/ai-openapi.yaml
openapi: 3.1.0
info:
  title: SysadminAnywhere REST API
  version: 5.0.0
  description: >-
    Examples for the implemented Directory, Incident, and Inventory service APIs.
    All operations except authentication require an administrator JWT obtained from
    POST /api/ldap/authenticate. See api-reference-ru.md and api-reference-en.md
    for service URLs and usage notes.
servers:
  - url: http://localhost:8081
    description: Directory service (local development)
  - url: http://localhost:8082
    description: Inventory service (local development)
  - url: http://localhost:8083
    description: Incident service (local development)
security:
  - bearerAuth: []
paths:
...
  /api/users:
    get:
      servers:
        - url: http://localhost:8081
      tags: [Users]
      summary: Get a page of users using an LDAP filter
      operationId: getUsers
      parameters:
        - name: page
          in: query
          schema: { type: integer, default: 0, minimum: 0 }
        - name: size
          in: query
          schema: { type: integer, default: 10, minimum: 1 }
        - name: sort
          in: query
          description: Accepted by the endpoint for compatibility; currently not applied.
          schema: { type: string, default: '' }
        - name: filters
          in: query
          required: true
          description: LDAP filter expression.
          schema: { type: string }
          example: (objectClass=user)
        - name: attributes
          in: query
          required: true
          description: Repeat this parameter for each LDAP attribute to return.
          style: form
          explode: true
          schema:
            type: array
            items: { type: string }
          example: [cn, mail, sAMAccountName]
      responses:
        '200':
          description: Page of users
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserPage'
        '400':
          description: Invalid LDAP filter
    post:
      servers:
        - url: http://localhost:8081
      tags: [Users]
      summary: Create an Active Directory user
      operationId: createUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddUserRequest'
            example:
              distinguishedName: CN=Alex Morgan,OU=Users,DC=example,DC=com
              cn: Alex Morgan
              displayName: Alex Morgan
              firstName: Alex
              lastName: Morgan
              initials: AM
              password: temporary-password
              cannotChangePassword: false
              passwordNeverExpires: false
              accountDisabled: true
              mustChangePassword: true
      responses:
        '201':
          description: User created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DirectoryEntry'
        '400':
          description: Missing required user fields or invalid data
  /api/users/bulk/change-status:
    post:
      servers:
        - url: http://localhost:8081
      tags: [Users, Bulk operations]
      summary: Enable or disable multiple user accounts
      operationId: bulkChangeUserStatus
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BulkUserStatusRequest'
            example:
              distinguishedNames:
                - CN=Alex Morgan,OU=Users,DC=example,DC=com
                - CN=Sam Lee,OU=Users,DC=example,DC=com
              accountDisabled: true
      responses:
        '200':
          description: Per-object operation results
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BulkOperationResult'
        '400':
          description: No distinguished names were supplied
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    AddUserRequest:
      type: object
      required: [distinguishedName, cn, password]
      properties:
        distinguishedName: { type: string }
        cn: { type: string }
        displayName: { type: string }
        firstName: { type: string }
        lastName: { type: string }
        initials: { type: string }
        password: { type: string, format: password }
        cannotChangePassword: { type: boolean, default: false }
        passwordNeverExpires: { type: boolean, default: false }
        accountDisabled: { type: boolean, default: false }
        mustChangePassword: { type: boolean, default: false }
    BulkUserStatusRequest:
      type: object
      required: [distinguishedNames]
      properties:
        distinguishedNames:
          type: array
          minItems: 1
          items: { type: string }
        accountDisabled: { type: boolean, description: 'true disables accounts; false enables them.' }
...
    DirectoryEntry:
      type: object
      description: Directory attributes are returned according to the requested attributes and object type.
      additionalProperties: true
      properties:
        distinguishedName: { type: string }
        cn: { type: string }
        displayName: { type: string }
        samAccountName: { type: string }
        emailAddress: { type: string }
    UserPage:
      type: object
      properties:
        content:
          type: array
          items: { $ref: '#/components/schemas/DirectoryEntry' }
        page: { type: integer }
        size: { type: integer }
        totalElements: { type: integer, format: int64 }
        totalPages: { type: integer }
    BulkOperationResult:
      type: object
      description: Per-object success and failure details returned by bulk operations.
      properties:
        updated: { type: integer }
        failures:
          type: array
          items: { type: string }
...